Move types into proposals directory
